refactor(web): refine hooks naming strategy
